requirements.txt
flask==3.0.2
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
numpy==1.26.4
shap==0.44.1
scikit-image==0.22.0
matplotlib==3.8.3
scikit-learn==1.4.0
tqdm
seaborn
ipykernel
notebook